VMO-6851/VMO-7043 re-emit 'change' event and handle it in top ResourceViewer component (avoid auto-save in block editor from ResourceEditorCell)
