+++ src/components/interaction-designer/resource-viewer/index.ts
@@ -1,3 +1,1 @@
 export {default as ResourceViewerBlock} from './ResourceViewerBlock.vue'
-
-export const DEBOUNCE_FLOW_PERSIST_MS = 1500
